chore: deleted unnecessary files and moved existing ones
