+++ my-vue-app/src/assets/styles/components/features-card.scss
@@ -1,4 +1,4 @@
-@import '../main/index';
+@import './main/index';
 
 .card {
   display: flex;
