--- my-vue-app/src/assets/styles/components/main-banner/main-banner.scss
@@ -1,112 +1,0 @@
-@import '../main/index';
-
-.main-banner {
-  position: relative;
-  padding: 40px;
-  background: $gradient-main-bg;
-
-  &__content {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    text-align: center;
-  }
-
-  &__ellipse-top {
-    position: absolute;
-    top: 20%;
-    left: 52%;
-    transform: translate(-54%, -80%);
-
-    @media (max-width: $breakpoint-xxl) {
-      transform: translate(-100%, -100%);
-    }
-  }
-
-  &__title {
-    margin-top: 20px;
-  }
-
-  &__text-block {
-    flex-direction: column;
-    width: 436px;
-
-    @media (max-width: $breakpoint-md) {
-      width: 100%;
-      margin-top: 32px;
-      padding-right: 16px;
-      gap: 20px;
-    }
-  }
-
-  &__description {
-    margin-top: -20px;
-
-    @media (max-width: $breakpoint-md) {
-      font-size: 16px;
-    }
-  }
-
-  &__secondary-title {
-    margin: 120px 0 32px;
-
-    @media (max-width: $breakpoint-md) {
-      padding-right: 16px;
-      font-size: 40px;
-    }
-  }
-
-  &__button {
-    box-shadow: $button-shadow;
-  }
-
-  &__screen {
-    z-index: $ui-index-1;
-    width: 80%;
-    max-width: 1100px;
-    height: auto;
-    margin-top: 40px;
-    border-radius: 8px;
-    box-shadow: 0 20px 40px $color-shadow-base;
-    aspect-ratio: 1100 / 616;
-
-    @media (max-width: $breakpoint-md) {
-      width: calc(100% - 40px);
-      margin-top: 32px;
-    }
-  }
-
-  &__cards-wrapper {
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    gap: 24px;
-    z-index: $ui-index-2;
-
-    @media (max-width: $breakpoint-md) {
-      grid-template-columns: 1fr;
-      gap: 16px;
-    }
-
-    &-ellipse-top {
-      z-index: $ui-index-1;
-      transform: translate(42%, -100%);
-
-      @media (max-width: $breakpoint-xxl) {
-        transform: translate(-100%, -100%);
-      }
-    }
-
-    &-ellipse-down {
-      z-index: $ui-index-1;
-      transform: translate(-22%, -144%);
-
-      @media (max-width: $breakpoint-xxl) {
-        transform: translate(-100%, -100%);
-      }
-    }
-  }
-
-  &__down-banner {
-    margin-top: -1240px;
-  }
-}
